requirements.txt
# AI Job Screener - Python Dependencies
# Compatible with Python 3.12

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# OpenAI API
openai==1.10.0

# Vector Database & Search
faiss-cpu==1.8.0
numpy==1.26.4

# Data Processing
pandas==2.2.0
requests==2.31.0

# Additional utilities
pydantic==2.6.0
python-multipart==0.0.7

# For async support
httpx==0.26.0
aiofiles==23.2.1